Remove deprecated modules and update embedder API
