+++ pycontextify/embedder_factory.py
@@ -7,7 +7,7 @@ making it easy to switch between different providers and add new ones.
 import logging
 from typing import Any, Dict, List, Type
 
-from .embedder_base import BaseEmbedder, EmbeddingError, ProviderNotAvailableError
+from .embedder import BaseEmbedder, EmbeddingError, ProviderNotAvailableError
 from .embedder_sentence_transformers_embedder import SentenceTransformersEmbedder
 
 logger = logging.getLogger(__name__)
