--- pycontextify/indexer.py
@@ -1,23 +1,0 @@
-"""Indexing runtime for PyContextify.
-
-This module now serves as the single entry point for index related
-configuration and orchestration utilities.
-"""
-
-from .index_codebase import CodeLoader, CodebaseIndexer
-from .index_document import DocumentIndexer, DocumentLoader, PDFLoader
-from .index_webpage import WebpageIndexer, WebpageLoader
-from .indexer_manager import IndexManager
-from .config import Config
-
-__all__ = [
-    "Config",
-    "IndexManager",
-    "CodeLoader",
-    "CodebaseIndexer",
-    "DocumentLoader",
-    "DocumentIndexer",
-    "PDFLoader",
-    "WebpageLoader",
-    "WebpageIndexer",
-]
