--- pycontextify/indexer_pdf_loader.py
@@ -1,5 +1,0 @@
-"""Compatibility shim for :mod:`pycontextify.index_document`."""
-
-from .index_document import PDFLoader
-
-__all__ = ["PDFLoader"]
